Move testcases to the existing test file.
